More examples of unique_ptr
This code shows the destructor of a class being called by unique_ptr leaving contex, including by exception handler.
It also shows that when a function is called passing a unique_ptr as argument, it moves ownership.
In Part 3 is possible see that the context of a unique Ptr is the local of the uniquePtr variable is declared, not where the referenced content is constructed.
E.g:
	{
	std::unique_ptr<Y> y3; <-- Declaration of unique_ptr, this is the context for destruction.
		{
		y3 = std::make_unique<Y>(3); <-- Y Constructor is called here. But the pointer was already declared in other context.
		} <-- NO destructor
	}<-- This is where the destructor will be called.